+++ pkg/utils/kubernetes/resources/hash.go
@@ -4,6 +4,9 @@ import (
 	"fmt"
 
 	"github.com/gohugoio/hashstructure"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/kong/gateway-operator/pkg/consts"
 )
 
 // CalculateHash calculates the hash of the given object.
@@ -17,4 +20,21 @@ func CalculateHash[T any](
 	}
 
 	return fmt.Sprintf("%0x", hash), nil
+}
+
+// SpecHashMatchesAnnotation calculates the hash of the given spec and returns boolean
+// indicating whether the hash matches the one in the annotations of the given
+// object.
+func SpecHashMatchesAnnotation[T any](
+	spec T,
+	obj client.Object,
+) (bool, error) {
+	hash, err := CalculateHash(spec)
+	if err != nil {
+		return false, fmt.Errorf("failed to calculate hash spec from %T: %w", spec, err)
+	}
+	if h, ok := obj.GetAnnotations()[consts.AnnotationSpecHash]; !ok || h != hash {
+		return false, nil
+	}
+	return true, nil
 }